Create a simple Jest-based integration test.
Usage: `yarn test`.

The test currently walks through the motions of adding lamports to a
payer account, minting a couple tokens, token accounts, and transferring
tokens around.

Next will be creating the logic to actually generate a Cove vault,
deploying the Cove logic to devnet, and storing tokens in a HODL vault.